+++ lib/mobile/home_page.dart
@@ -219,7 +219,7 @@ class _HomePageState extends State<HomePage> with ChatUIKitThemeMixin {
       decoration: BoxDecoration(
         boxShadow: [
           BoxShadow(
-            color: Colors.black.withOpacity(0.2),
+            color: Colors.black.withValues(alpha: 0.2),
             blurRadius: 10,
             offset: const Offset(0, -5),
           ),