+++ lib/uikit/lib/chat_uikit/src/ui/views/group_details_view.dart
@@ -511,7 +511,7 @@ class _GroupDetailsViewState extends State<GroupDetailsView>
           builder: (context, value, child) {
             return CupertinoSwitch(
               value: isNotDisturb.value,
-              activeColor: theme.color.isDark
+              activeTrackColor: theme.color.isDark
                   ? theme.color.primaryColor6
                   : theme.color.primaryColor5,
               trackColor: theme.color.isDark